refactor: correção calculo

+++ api-calculo/Domain/Interfaces/ICalculaJurosService.cs
@@ -4,6 +4,6 @@ namespace api_calculo.Domain.Interfaces
 {
     public interface ICalculaJurosService
     {
-        Task<decimal> JurosCompostos(int valorinicial, int meses);
+        Task<decimal> JurosCompostos(double valorinicial, int meses);
     }
 }
